requirements.txt
Python-telegram-bot==13.9.0
requests
ping3
ipinfo